+++ node/src/types/kernel/api/asset/upload/response.d.ts
@@ -15,21 +15,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+/**
+ * Upload assets
+ */
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    data: IData;
+    /**
+     * status message
+     */
+    msg: string;
+}
 
-/* directories */
-export * as asset from "./asset";
-export * as attr from "./attr";
-export * as block from "./block";
-export * as convert from "./convert";
-export * as export from "./export";
-export * as file from "./file";
-export * as filetree from "./filetree";
-export * as notebook from "./notebook";
-export * as notification from "./notification";
-export * as query from "./query";
-export * as search from "./search";
-export * as storage from "./storage";
-export * as system from "./system";
-export * as template from "./template";
-
-/* files */
+/**
+ * the result of uploading assets
+ */
+export interface IData {
+    errFiles: string[] | null;
+    succMap:  { [key: string]: string };
+}